Adding server side rendering for a list of products.
+Moving firebase to services directory
+Pulling in products from database in /listing
+Initialize custom elelements at import
+Add launcher for js for /listing route
+Add styles for serverside rendered pie card
+Add example markup for narrow pie card
