+++ .gitignore
@@ -1,4 +1,4 @@
 node_modules
 .DS_Store
 dist
-keys
+src/data/keys
