+++ package.json
@@ -57,7 +57,7 @@
     "test:node": "./scripts/test-node.sh",
     "test:static": "./node_modules/.bin/karma start",
     "test": "npm run test:node && npm run test:static",
-    "tools:piegen": "./node_modules/.bin/babel-node tools/piegen/firebase.js"
+    "tools:piegen": "./node_modules/.bin/babel-node tools/piegen/generate.js"
   },
   "devDependencies": {
     "karma": "^1.7.1",
